web/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Redmail==0.3.0
Flask-Babel==3.1.0
usfm-grammar==3.0.0a5
wildebeest-nlp==0.9.2
sil-machine==0.9.3
